+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `isSerializable` function which allows to check if target is an instance of serializable class
 * Added class name to serialization/deserialization error message
 
 v1.2.3